Time: 6 ms (44.61%) | Memory: 43.2 MB (68.44%) - LeetSync

+++ 242-valid-anagram/valid-anagram.java
@@ -2,7 +2,7 @@ class Solution {
     public boolean isAnagram(String s, String t) {
         if (s.length() != t.length())
             return false;
-        int[] chars = new int[26]; 
+        short[] chars = new short[26]; 
         for (int i = 0; i < s.length();i++){
             chars[s.charAt(i) - 'a'] += 1;
             chars[t.charAt(i) - 'a'] -= 1;
